requirements.txt
fastapi
uvicorn[standard]
requests
sentence-transformers
rapidfuzz
python-dotenv

